+++ lib/laveno/board.ex
@@ -23,12 +23,8 @@ defmodule Laveno.Board do
           moves: list(bitstring())
         }
 
-  use Bitwise
   require Logger
 
-  @offset_row 49
-  @offset_column 97
-  @pieces_set [:P, :p, :N, :n, :B, :b, :K, :k, :Q, :q, :R, :r]
   @w_pieces [:P, :R, :N, :B, :K, :Q]
   @b_pieces [:p, :r, :n, :b, :k, :q]
 
@@ -179,10 +175,12 @@ defmodule Laveno.Board do
          piece <- Utils.which_piece?(board, from_square),
          true <- right_turn?(board, piece) do
       is_capture = Utils.which_piece?(board, to_square) != nil
+      is_ep = piece in [:P, :p] and board.en_passant == to_square
       is_pawn_move = piece in [:P, :p]
 
       board
       |> update_castling_rights(piece, from_square, to_square)
+      |> then(fn b -> if is_ep, do: clear_square(b, <<c2, r1>>), else: b end)
       |> clear_square(from_square)
       |> clear_square(to_square)
       |> place_piece(piece, to_square)
@@ -194,6 +192,37 @@ defmodule Laveno.Board do
       _ -> {:error, "invalid move"}
     end
   end
+
+  @doc """
+  Apply a movegen move without `valid_move?` or move-history logging.
+  Search uses this; UCI position still goes through `move/2`.
+  """
+  def apply_search(board = %__MODULE__{}, <<c1::8, r1::8, c2::8, r2::8, _::binary>> = move) do
+    from = <<c1, r1>>
+    to = <<c2, r2>>
+    piece = Utils.which_piece?(board, from)
+
+    if piece == nil do
+      {:error, "invalid move"}
+    else
+      is_capture =
+        Utils.which_piece?(board, to) != nil or
+          (piece in [:P, :p] and board.en_passant == to)
+
+      board
+      |> update_castling_rights(piece, from, to)
+      |> Map.put(:bb, Utils.apply_pseudo(board, move))
+      |> proc_en_passant(piece, move)
+      |> then(
+        if piece in [:P, :p] or is_capture,
+          do: &reset_halfmove_clock/1,
+          else: &increment_count/1
+      )
+      |> flip_active_color()
+    end
+  end
+
+  def apply_search(_, _), do: {:error, "invalid move"}
 
   def right_turn?(%{active_color: <<0::1>>}, piece)
       when piece in @w_pieces,
@@ -272,7 +301,7 @@ defmodule Laveno.Board do
   def proc_en_passant(
         board,
         pawn,
-        move = <<c1::8, r1::8, c2::8, r2::8>>
+        <<_c1::8, r1::8, c2::8, r2::8>>
       )
       when pawn in [:P, :p] do
     case abs(r2 - r1) == 2 do