+++ test/fen_test.exs
@@ -10,7 +10,7 @@ defmodule LavenoTest.FenTest do
     test "FEN initial position" do
       fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1"
 
-      {fen_state, board} = fen = Fen.load(fen)
+      {_fen_state, board} = Fen.load(fen)
 
       Render.print_board(board)
 
@@ -22,7 +22,7 @@ defmodule LavenoTest.FenTest do
     test "FEN initial positionm after 1.e4" do
       fen = "rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq e3 0 1"
 
-      {fen_state, board} = fen = Fen.load(fen)
+      {_fen_state, board} = Fen.load(fen)
 
       Render.print_board(board)
 
@@ -34,7 +34,7 @@ defmodule LavenoTest.FenTest do
     test "FEN initial positionm after 1...c5" do
       fen = "rnbqkbnr/pp1ppppp/8/2p5/4P3/8/PPPP1PPP/RNBQKBNR w KQkq c6 0 2"
 
-      {fen_state, board} = fen = Fen.load(fen)
+      {_fen_state, board} = Fen.load(fen)
 
       Render.print_board(board)
 
@@ -46,7 +46,7 @@ defmodule LavenoTest.FenTest do
     test "FEN initial positionm after 2.Nf3" do
       fen = "rnbqkbnr/pp1ppppp/8/2p5/4P3/5N2/PPPP1PPP/RNBQKB1R b KQkq - 1 2"
 
-      {fen_state, board} = fen = Fen.load(fen)
+      {_fen_state, board} = Fen.load(fen)
 
       Render.print_board(board)
 
@@ -73,7 +73,7 @@ defmodule LavenoTest.FenTest do
     test "FEN just a tactic position. Nd6#" do
       fen = "1b1B1Nb1/rRqp1kPp/1p5P/8/2N2Pp1/p3R3/B3Q3/2K5"
 
-      {fen_state, board} = fen = Fen.load(fen)
+      {_fen_state, board} = Fen.load(fen)
 
       Render.print_board(board)
     end